# Conflicts:
#	src/main/java/com/massil/config/JobRunrConfig.java
#	src/main/java/com/massil/dto/AppraisalVehicleCard.java
#	src/main/java/com/massil/persistence/mapper/AppraisalVehicleMapper.java
#	src/main/java/com/massil/persistence/mapper/OffersMapper.java
#	src/main/java/com/massil/repository/elasticRepo/AppraisalVehicleERepo.java
#	src/main/java/com/massil/services/ConfigCodesService.java
#	src/main/java/com/massil/services/impl/ConfigCodesServiceImpl.java

+++ src/main/java/com/massil/dto/ConfigDropDown.java
@@ -39,5 +39,5 @@ public class ConfigDropDown {
     @NotNull
     @Size(max = 20,message = "lenght must be less than equal to 20 characters")
     private String configGroup;
-    private Integer intValue;
+    private Double intValue;
 }
